seperated ProjectSideBar into two diffren't components
