+++ src/App.scss
@@ -8,6 +8,7 @@ body {
   font-family: acumin-pro, "Helvetica Neue", Helvetica, Arial; 
   font-smooth: always;
   -webkit-font-smoothing: antialiased;
+  overflow-y: hidden;
 }
 
 .App {
