+++ src/components/content/project_components/Controllers/Controllers.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import ProjectSetupSidebar from '../ProjectSetupSidebar/ProjectSetupSidebar';
+import ProjectSidebar from '../ProjectSetupSidebar/ProjectSidebar';
 import Header from '../../../Header/Header';
 import addIcon from '../../../../img/icons/add-icon.svg';
 import './controllers.scss';
@@ -131,7 +131,7 @@ class Controllers extends Component {
         <div>
             <Header />
             <div className="main-fix">
-                <ProjectSetupSidebar userid={userid} projectid={projectid}/>
+                <ProjectSidebar userid={userid} projectid={projectid}/>
                 <div className="controllers-container">
                     <div className="controllers-wrapper">
                         <div className="project-section-header"> Controllers </div>
