+++ src/components/content/project_components/Endpoints/Endpoints.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import Header from '../../../Header/Header';
-import ProjectSetupSidebar from '../ProjectSetupSidebar/ProjectSetupSidebar'
+import ProjectSidebar from '../ProjectSetupSidebar/ProjectSidebar'
 import './endpoints.scss';
 
 class Endpoints extends Component {
@@ -151,7 +151,7 @@ class Endpoints extends Component {
       <div>
       <Header />
         <div className="main-fix">
-          <ProjectSetupSidebar userid={userid} projectid={projectid}/>
+          <ProjectSidebar userid={userid} projectid={projectid}/>
           <div className="endpoints-container">
           
           <div className="endpoints-wrapper">
