+++ src/components/content/project_components/Features/features.scss
@@ -5,6 +5,37 @@
     background-color: rgba(0,0,0,0.5);
     height: calc(100vh - 50px);
     overflow-y: auto;
+
+    
+    &::-webkit-scrollbar {
+        width: 15px;
+        padding-right: 5px;
+        -webkit-overflow-scrolling: touch;
+    }
+    
+
+    &::-webkit-scrollbar-track {
+        margin-top: 3%;
+        margin-bottom: 3%;
+        
+        -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3); 
+        -webkit-border-radius: 10px;
+        border-radius: 10px;
+    }
+ 
+
+    &::-webkit-scrollbar-thumb {
+        -webkit-border-radius: 10px;
+        border-radius: 10px;
+        background-color: rgba(50, 50, 50, 0.80); 
+        -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.5);
+        
+    }
+    &::-webkit-scrollbar-thumb:window-inactive {
+        background-color: rgba(50, 50, 50, 0.40);       
+    }
+
+    
 }
 
 .container-wrapper {
@@ -13,7 +44,6 @@
     align-items: center;
     width: 100%;
     margin-bottom: 40px;
-    
 }
 
 .features-area {
