+++ src/components/content/project_components/ProjectSetupSidebar/BackgroundMenu/backgroundmenu.scss
@@ -1,0 +1,197 @@
+//CHANGE BACKGROUND MENU UI
+
+.project-background-header{
+    display: flex;
+    width: 100%;
+    height: 50px;
+    align-items: center;
+    border-bottom: 1px solid #DCDCDC;
+    padding-left: 0px;
+    padding-right: 0px;
+    
+    line-height: 56px;
+
+    .rename-project{
+        display: none;
+    }
+
+    
+
+    .change-background-url{
+        display: block;
+        padding-left: 24px;
+        padding-right: 24px;
+        text-align: left;
+        font-size: 22px;
+        color: black;
+        max-height: 100%;
+        outline: none;
+        width: 100%;
+        background: transparent;
+        border: none;
+    }
+    ::-webkit-input-placeholder {
+        text-align: left;
+        color: #595959;
+        font-size: 18px;
+        font-weight: 300;
+    }
+}
+
+
+
+.background-list{
+    align-items: center;
+    width: 100%;
+    display: flex;
+    text-align: left;
+    flex-direction: column;
+
+    height: calc(90vh - 100px);
+    padding-left: 0px;
+    overflow-y: scroll !important;
+
+    .nav-item{
+        display: none;
+    }
+
+    &::-webkit-scrollbar {
+        width: 10px;
+        padding-left: 5px;
+        height: 100%;            
+    }
+
+    &::-webkit-scrollbar-thumb {
+        -webkit-border-radius: 10px;
+        border-radius: 10px;
+        background-color: #C4C9CC;
+        position: absolute;
+        margin-left: 5px;
+           
+    }
+
+    &::-webkit-scrollbar-track {
+    margin-left: 5px;
+    }
+
+    input{
+        max-width: 80%;
+        border-radius: 10px;
+        min-height: 30px;
+        padding-left: 10px;
+        padding-right: 10px;
+        outline: none;
+        border: 1px solid #DCDCDC;
+    }
+
+    .no-background-box{
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        margin-top: 20px;
+        background: #DCDCDC;
+        border: 1px solid #DCDCDC;
+        min-height: 100px;
+        width:80%;
+        background-size: 100%;
+        background-position: 50% 50%;
+        border-radius: 10px; 
+    }
+
+    .background-box{
+        display: flex;
+        flex-direction: column;
+        justify-content: flex-end;
+        margin-top: 20px;
+        border: 1px solid #DCDCDC;
+        min-height: 100px;
+        width:80%;
+        background-size: 100%;
+        background-position: 50% 50%;
+        border-radius: 10px;        
+        cursor: pointer;
+
+        .attribution{
+            opacity: 0;
+            user-select: none;
+            transition: 0.2s ease-in-out;
+        }
+
+        &:hover{
+            .attribution{
+                opacity: 1;
+                align-items: center;
+                display: flex;
+                height: 25px;
+                width: 100%;
+                background-color: rgba(0, 0, 0, 0.5);
+                border-radius: 0px 0px 10px 10px;
+                transition: 0.2s ease-in-out;
+    
+                label{
+                    margin-left: 10px;
+                    cursor: pointer;
+                    color: hsla(0,0%,100%,.8);
+                    
+                    &:hover{
+                        text-decoration: underline;
+                    }
+                }
+            }
+        }
+    }
+}
+
+.project-background-footer{
+    display: block;
+    position: absolute;
+    top: calc(100% - 50px);
+    background: transparent;
+    height: 50px;
+    width: 192px;
+    padding-left: 24px;
+    padding-right: 24px;
+
+    .project-sidebar-buttonset{
+        display: flex;
+        align-items: center;
+
+        .changeBackgroundBtn{
+            height: 35px;
+            width: 100px;
+            margin-right: 5px;
+            border-radius: 3px;
+            outline: none;
+            border: none;
+            color: white;
+            font-weight: bold;
+            background: #0079BF;
+
+            &:active{
+                background: #006DAC;
+            }                
+        }
+
+        .closeBackgroundList{
+            width: 18px;
+            height: 18px;
+            background: transparent;
+            border: none;
+            margin-left: 5px;
+            cursor: pointer;
+
+            span{
+                display: block;
+                background-repeat: no-repeat;
+                background-image: url('../../../../../img/icons/tracker-close-x-icon.svg');
+                background-position: 100%;
+                height: 100%;
+                width: 100%;
+            }
+        }
+    }
+    label{
+        display: none;
+    }
+}
