+++ src/components/content/project_components/ProjectSetupSidebar/NavMenu/navmenu.scss
@@ -1,0 +1,100 @@
+
+    .project-sidebar-header{
+        display: flex;
+        max-width: 240px;
+        height: 50px;
+        border-bottom: 1px solid #DCDCDC;
+        line-height: 56px;
+
+        .change-background-url{
+            display: none;
+        }
+
+        .rename-project{
+            padding-left: 24px;
+            padding-right: 24px;
+            display: block;
+            text-align: left;
+            font-size: 22px;
+            color: black;
+            max-height: 100%;
+            outline: none;
+            width: 80%;
+            background: transparent;
+            border: none;
+            
+           
+        }
+        ::-webkit-input-placeholder {
+            text-align: left;
+            color: #595959;
+            font-size: 22px;
+            font-weight: 300;
+        }
+    }
+
+
+
+    .nav-list{
+        align-items: center;
+        min-width: 240px;
+        display: flex;
+        text-align: left;
+        flex-direction: column;
+        justify-content: space-between;
+        margin-top: 16px;
+        height: calc(90% - 50px);
+        padding-left: 0px;
+
+        .no-background-box{
+            display: none;
+        }
+
+        .background-box{
+            display: none;
+        }
+    }
+
+
+
+    .nav-item {
+
+        padding-left: 24px;
+        width: 240px;
+        line-height: 40px;
+        font-size: 16px;
+        list-style: none;
+        color: #8C8984;
+        text-decoration: none;
+        background: transparent;
+        &:hover{
+            background: rgba(66, 133, 244, 0.25);
+        }
+
+    }
+
+    .project-sidebar-footer{
+        display: block;
+        position: absolute;
+        top: calc(100% - 50px);
+        color: blue;
+        background: transparent;
+        height: 50px;
+        width: 240px;
+        padding: 0px;
+
+        .project-sidebar-buttonset{
+            display: none;
+        }
+
+        label{
+            display: block;
+            position: absolute;
+            padding-left: 24px;
+            padding-top: 16px;
+            color: #8C8984;
+            text-decoration: underline;
+            cursor: pointer;
+        }
+    }
+
