+++ src/components/content/project_components/ProjectSetupSidebar/projectsidebar.scss
@@ -1,0 +1,51 @@
+
+    .project-sidebar-container {
+        min-width: 240px;
+        height: calc(100vh - 50px);
+        background-color: white;
+        display: flex;
+        flex-direction: column;
+        align-items: flex-start;
+        border-right: 1px solid #DCDCDC;
+        overflow-x: hidden;
+    }
+
+.navMenu-wrapper{
+    height: 100%;
+    transform: translateX(0%);
+    opacity: 1;
+    transition: 0.2s ease-in-out;
+    
+}
+
+.navMenu--hide{
+    height: 100%;
+    transform: translateX(-100%);
+    opacity: 0.2;
+    position: absolute;
+}
+
+.backgroundMenu-wrapper{
+    height: 100%;
+    transform: translateX(0%);
+    opacity: 1;
+    max-width: 240px;
+    transition: 0.2s ease-in-out;
+    position: relative;
+}
+
+.backgroundMenu--hide{
+    height: 100%;
+    transform: translateX(-100%);
+    opacity: 0.2;
+    transition: 0.2s ease-in-out;
+    position: absolute;
+    overflow-x: hidden;
+}
+
+
+
+
+
+
+
