+++ src/components/content/project_components/Tracking/tracking.scss
@@ -65,8 +65,7 @@
             
         }
         &::-webkit-scrollbar-thumb:window-inactive {
-            background-color: rgba(50, 50, 50, 0.40); 
-            
+            background-color: rgba(50, 50, 50, 0.40);       
         }
     }
 }
